backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
watchfiles==0.21.0
httpx==0.25.2
PyYAML==6.0.1
asyncio-mqtt==0.16.1
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
jinja2==3.1.2
markdown==3.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6